feat: add verifications on postos

+++ VacinaAPI/PostoVacinacao/Handlers/DeletePostoVacinacaoHandler.cs
@@ -1,21 +1,23 @@
-﻿using AutoMapper;
-using Infrastructure.Repositories;
+﻿using Infrastructure.Repositories;
+using VacinaAPI.Exceptions;
 
 namespace VacinaAPI.PostoVacinacao.Handlers;
 
 public class DeletePostoVacinacaoHandler
 {
     private readonly IPostosRepository _postosRepository;
-    private readonly IMapper _mapper;
     
-    public DeletePostoVacinacaoHandler(IPostosRepository postosRepository, IMapper mapper)
+    public DeletePostoVacinacaoHandler(IPostosRepository postosRepository)
     {
         _postosRepository = postosRepository;
-        _mapper = mapper;
     }
     
     public async Task<int?> DeletePostoById(int id)
     {
+        bool hasVacinas = await _postosRepository.HasVacinasRelatedById(id);
+        if (hasVacinas)
+            throw new CantDeleteRelationException("Posto", "Vacina");
+        
         int? deletedId = await _postosRepository.DeletePostoById(id);
         await _postosRepository.FlushChanges();
         
@@ -24,6 +26,10 @@ public class DeletePostoVacinacaoHandler
     
     public async Task<int?> DeletePostoByName(string name)
     {
+        bool hasVacinas = await _postosRepository.HasVacinasRelatedByName(name);
+        if (hasVacinas)
+            throw new CantDeleteRelationException("Posto", "Vacina");
+        
         int? deletedId = await _postosRepository.DeletePostoByName(name);
         await _postosRepository.FlushChanges();
